+++ packages/origin-backend/src/app.module.ts
@@ -16,6 +16,7 @@ import { OrganizationModule } from './pods/organization/organization.module';
 import { UserModule } from './pods/user/user.module';
 import { EmailConfirmationModule } from './pods/email-confirmation/email-confirmation.module';
 import { providers } from '.';
+import { CertificationRequestModule } from './pods/certification-request/certification-request.module';
 
 const ENV_FILE_PATH = path.resolve(__dirname, '../../../../../.env');
 
@@ -36,7 +37,8 @@ export class AppModule {
                 OrganizationModule,
                 DeviceModule.register(smartMeterReadingsAdapter),
                 AuthModule,
-                CertificateModule.register(smartMeterReadingsAdapter),
+                CertificateModule,
+                CertificationRequestModule.register(smartMeterReadingsAdapter),
                 AdminModule,
                 EmailConfirmationModule
             ],
