+++ packages/origin-backend/test/device.e2e-spec.ts
@@ -17,7 +17,7 @@ import { bootstrapTestInstance, registerAndLogin } from './origin-backend';
 import { DeviceService } from '../src/pods/device/device.service';
 import { OrganizationService } from '../src/pods/organization/organization.service';
 import { UserService } from '../src/pods/user/user.service';
-import { CertificationRequestService } from '../src/pods/certificate/certification-request.service';
+import { CertificationRequestService } from '../src/pods/certification-request/certification-request.service';
 import { DatabaseService } from './database.service';
 
 describe('Device e2e tests', () => {
